Bugs: dependency issues with scripts
modulated some more code off eventlistners.
