+++ eventListeners.js
@@ -1,10 +1,109 @@
-import { showDeleteConfirmationModal, showAddStationModal, showAssignAssociateModal, showRemoveAssociateModal, showWarningModal } from './modals.js';
+import { initializeAreaButtons } from './areaButtons.js';
 
 document.addEventListener('DOMContentLoaded', function() {
     const container = document.getElementById('container');
     let areasData = JSON.parse(localStorage.getItem('areasData')) || [];
     let rosterData = JSON.parse(localStorage.getItem('rosterData')) || [];
     const colors = ['bg-dark-blue', 'bg-dark-cyan', 'bg-dark-teal', 'bg-dark-navy', 'bg-dark-slate', 'bg-dark-steel', 'bg-dark-azure', 'bg-dark-indigo'];
+
+    function showDeleteConfirmationModal(message, onDelete) {
+        const deleteModal = new bootstrap.Modal(document.getElementById('deleteConfirmationModal'));
+        document.getElementById('deleteConfirmationMessage').textContent = message;
+        document.getElementById('confirmDeleteButton').onclick = function() {
+            onDelete();
+            deleteModal.hide(); // Close the modal after deletion
+        };
+        deleteModal.show();
+    }
+
+    function showAddStationModal(onAdd) {
+        const addStationModal = new bootstrap.Modal(document.getElementById('addStationModal'));
+        const stationNameInput = document.getElementById('stationNameInput');
+        stationNameInput.value = ''; // Clear the input field
+        document.getElementById('confirmAddStationButton').onclick = function() {
+            const stationName = stationNameInput.value.trim();
+            if (stationName) {
+                onAdd(stationName);
+                addStationModal.hide(); // Close the modal after adding
+            } else {
+                alert('Please enter a station name.');
+            }
+        };
+        document.getElementById('addStationModal').addEventListener('hidden.bs.modal', function () {
+            document.body.classList.remove('modal-open');
+            document.querySelector('.modal-backdrop').remove();
+        });
+        addStationModal.show();
+    }
+
+    function showAssignAssociateModal(stationIndex, areaIndex) {
+        const assignAssociateModal = new bootstrap.Modal(document.getElementById('assignAssociateModal'));
+        const associateSelect = document.getElementById('associateSelect');
+        associateSelect.innerHTML = ''; // Clear the select options
+
+        areasData[areaIndex].associates.forEach((associate, index) => {
+            const option = document.createElement('option');
+            option.value = index;
+            option.textContent = associate.name;
+            associateSelect.appendChild(option);
+        });
+
+        document.getElementById('confirmAssignAssociateButton').onclick = function() {
+            const selectedAssociateIndex = associateSelect.value;
+            if (selectedAssociateIndex !== '') {
+                const associate = areasData[areaIndex].associates[selectedAssociateIndex];
+                areasData[areaIndex].stations[stationIndex].associates.push(associate);
+                updateAreasData();
+                renderAreas();
+                assignAssociateModal.hide();
+            } else {
+                alert('Please select an associate.');
+            }
+        };
+
+        assignAssociateModal.show();
+    }
+
+    function showRemoveAssociateModal(stationIndex, areaIndex) {
+        const removeAssociateModal = new bootstrap.Modal(document.getElementById('removeAssociateModal'));
+        const associateSelect = document.getElementById('removeAssociateSelect');
+        associateSelect.innerHTML = ''; // Clear the select options
+
+        areasData[areaIndex].stations[stationIndex].associates.forEach((associate, index) => {
+            const option = document.createElement('option');
+            option.value = index;
+            option.textContent = associate.name;
+            associateSelect.appendChild(option);
+        });
+
+        document.getElementById('confirmRemoveAssociateButton').onclick = function() {
+            const selectedAssociateIndex = associateSelect.value;
+            if (selectedAssociateIndex !== '') {
+                areasData[areaIndex].stations[stationIndex].associates.splice(selectedAssociateIndex, 1);
+                updateAreasData();
+                renderAreas();
+                removeAssociateModal.hide();
+            } else {
+                alert('Please select an associate.');
+            }
+        };
+
+        removeAssociateModal.show();
+    }
+
+    function showWarningModal(message, onConfirm, onCancel) {
+        const warningModal = new bootstrap.Modal(document.getElementById('warningModal'));
+        document.getElementById('warningMessage').textContent = message;
+        document.getElementById('confirmWarningButton').onclick = function() {
+            onConfirm();
+            warningModal.hide(); // Close the modal after confirmation
+        };
+        document.getElementById('cancelWarningButton').onclick = function() {
+            onCancel();
+            warningModal.hide(); // Close the modal after cancellation
+        };
+        warningModal.show();
+    }
 
     // Function to render areas
     function renderAreas() {
@@ -47,226 +146,8 @@ document.addEventListener('DOMContentLoaded', function() {
                 return new bootstrap.Tooltip(tooltipTriggerEl);
             });
 
-            // Add event listener to the plus sign button
-            const newAAButton = newArea.querySelector('.newAAButton');
-            newAAButton.addEventListener('click', function() {
-                const badgeNumber = prompt("Badge Number?");
-                if (badgeNumber !== null) {
-                    const trimmedBadgeNumber = badgeNumber.trim();
-                    const matchedEntry = rosterData.find(entry => String(entry.badgeNumber).trim() === trimmedBadgeNumber);
-
-                    let login, name;
-                    if (matchedEntry) {
-                        login = matchedEntry.login;
-                        name = matchedEntry.name;
-                        addAssociate(trimmedBadgeNumber, login, name, index, areaContent);
-                    } else {
-                        document.getElementById('badgeNumberInput').value = trimmedBadgeNumber;
-                        const associateModal = new bootstrap.Modal(document.getElementById('associateModal'));
-                        associateModal.show();
-
-                        document.getElementById('saveAssociateButton').addEventListener('click', function() {
-                            login = document.getElementById('loginInput').value.trim();
-                            name = document.getElementById('nameInput').value.trim();
-
-                            if (login && name) {
-                                addAssociate(trimmedBadgeNumber, login, name, index, areaContent);
-                                // Update roster data
-                                rosterData.push({ badgeNumber: trimmedBadgeNumber, login, name });
-                                localStorage.setItem('rosterData', JSON.stringify(rosterData));
-                                associateModal.hide();
-                                // Clear input fields after saving
-                                document.getElementById('badgeNumberInput').value = '';
-                                document.getElementById('loginInput').value = '';
-                                document.getElementById('nameInput').value = '';
-                            } else {
-                                alert('Please fill in both Login and Name fields.');
-                            }
-                        }, { once: true });
-                    }
-                }
-            });
-
-            // Add drag and drop event listeners to the plus sign button
-            newAAButton.addEventListener('dragover', function(e) {
-                e.preventDefault();
-                e.dataTransfer.dropEffect = 'move';
-                this.classList.add('drag-over');
-            });
-
-            newAAButton.addEventListener('drop', function(e) {
-                handleAssociateDrop(e, index, areaContent);
-            });
-
-            // Add event listeners to the new action buttons
-            const actionButton1 = newArea.querySelector('.actionButton1');
-            const actionButton2 = newArea.querySelector('.actionButton2');
-            actionButton1.addEventListener('click', function() {
-                const settingsModal = new bootstrap.Modal(document.getElementById('settingsModal'));
-                const areaIndex = this.getAttribute('data-index');
-                const area = areasData[areaIndex];
-                const stationsToggle = document.getElementById('stationsToggle');
-                const stationsList = document.getElementById('stationsList');
-                const addStationButton = document.getElementById('addStationButton');
-
-                stationsToggle.checked = area.stations && area.stations.length > 0;
-                stationsList.innerHTML = '';
-                if (stationsToggle.checked) {
-                    area.stations.forEach((station, stationIndex) => {
-                        const stationItem = document.createElement('li');
-                        stationItem.className = 'list-group-item d-flex justify-content-between align-items-center';
-                        stationItem.innerHTML = `
-                            <div class="me-2 p-2 ${colors[stationIndex % colors.length]}"></div>
-                            <span class="fw-bold">${station}</span>
-                            <div>
-                                <button class="btn btn-light assignAssociateButton" data-area-index="${areaIndex}" data-station-index="${stationIndex}" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Assign associate"><i class="bi bi-plus"></i></button>
-                                <button class="btn btn-light removeAssociateButton" data-area-index="${areaIndex}" data-station-index="${stationIndex}" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Remove associate"><i class="bi bi-dash"></i></button>
-                                <i class="bi bi-trash remove-station-icon" data-area-index="${areaIndex}" data-station-index="${stationIndex}"></i>
-                            </div>
-                        `;
-                        stationsList.appendChild(stationItem);
-
-                        // Add event listener to the assign associate button
-                        stationItem.querySelector('.assignAssociateButton').addEventListener('click', function() {
-                            const stationIndex = this.getAttribute('data-station-index');
-                            const areaIndex = this.getAttribute('data-area-index');
-                            showAssignAssociateModal(stationIndex, areaIndex);
-                        });
-
-                        // Add event listener to the remove associate button
-                        stationItem.querySelector('.removeAssociateButton').addEventListener('click', function() {
-                            const stationIndex = this.getAttribute('data-station-index');
-                            const areaIndex = this.getAttribute('data-area-index');
-                            showRemoveAssociateModal(stationIndex, areaIndex);
-                        });
-
-                        // Add event listener to the remove station icon
-                        stationItem.querySelector('.remove-station-icon').addEventListener('click', function() {
-                            const stationIndex = this.getAttribute('data-station-index');
-                            area.stations.splice(stationIndex, 1);
-                            updateAreasData();
-                            renderAreas();
-                            this.parentElement.remove();
-                        });
-                    });
-                }
-
-                stationsToggle.addEventListener('change', function() {
-                    if (!this.checked) {
-                        showWarningModal('Turning off stations will delete all stations. Are you sure?', () => {
-                            stationsToggle.checked = false;
-                            addStationButton.style.display = 'none';
-                            area.stations = [];
-                            updateAreasData();
-                            stationsList.innerHTML = '';
-                        }, () => {
-                            stationsToggle.checked = true;
-                        });
-                    } else {
-                        addStationButton.style.display = 'inline-block';
-                    }
-                });
-
-                addStationButton.addEventListener('click', function() {
-                    showAddStationModal(function(stationName) {
-                        area.stations.push({ name: stationName, associates: [] });
-                        updateAreasData();
-                        const stationItem = document.createElement('li');
-                        stationItem.className = 'list-group-item d-flex justify-content-between align-items-center';
-                        stationItem.innerHTML = `
-                            <div class="me-2 p-2 ${colors[area.stations.length - 1 % colors.length]}"></div>
-                            <span class="fw-bold">${stationName}</span>
-                            <div>
-                                <button class="btn btn-light assignAssociateButton" data-area-index="${areaIndex}" data-station-index="${area.stations.length - 1}" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Assign associate"><i class="bi bi-plus"></i></button>
-                                <button class="btn btn-light removeAssociateButton" data-area-index="${areaIndex}" data-station-index="${area.stations.length - 1}" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Remove associate"><i class="bi bi-dash"></i></button>
-                                <i class="bi bi-trash remove-station-icon" data-area-index="${areaIndex}" data-station-index="${area.stations.length - 1}"></i>
-                            </div>
-                        `;
-                        stationsList.appendChild(stationItem);
-
-                        // Add event listener to the assign associate button
-                        stationItem.querySelector('.assignAssociateButton').addEventListener('click', function() {
-                            const stationIndex = this.getAttribute('data-station-index');
-                            const areaIndex = this.getAttribute('data-area-index');
-                            showAssignAssociateModal(stationIndex, areaIndex);
-                        });
-
-                        // Add event listener to the remove associate button
-                        stationItem.querySelector('.removeAssociateButton').addEventListener('click', function() {
-                            const stationIndex = this.getAttribute('data-station-index');
-                            const areaIndex = this.getAttribute('data-area-index');
-                            showRemoveAssociateModal(stationIndex, areaIndex);
-                        });
-
-                        // Add event listener to the remove station icon
-                        stationItem.querySelector('.remove-station-icon').addEventListener('click', function() {
-                            const stationIndex = this.getAttribute('data-station-index');
-                            area.stations.splice(stationIndex, 1);
-                            updateAreasData();
-                            renderAreas();
-                            this.parentElement.remove();
-                        });
-                    });
-                });
-
-                document.querySelectorAll('.remove-station-icon').forEach(icon => {
-                    icon.addEventListener('click', function() {
-                        const areaIndex = this.getAttribute('data-area-index');
-                        const stationIndex = this.getAttribute('data-station-index');
-                        areasData[areaIndex].stations.splice(stationIndex, 1);
-                        updateAreasData();
-                        renderAreas();
-                        this.parentElement.remove();
-                    });
-                });
-
-                document.getElementById('saveSettingsButton').addEventListener('click', function() {
-                    updateAreasData();
-                    settingsModal.hide();
-                });
-
-                settingsModal.show();
-            });
-            actionButton2.addEventListener('click', function() {
-                showDeleteConfirmationModal('Are you sure you want to delete this area?', () => {
-                    deleteArea(index);
-                });
-            });
-
-            // Add event listener to the remove associates button
-            const removeAssociatesButton = newArea.querySelector('.removeAssociatesButton');
-            removeAssociatesButton.addEventListener('click', function() {
-                const areaIndex = this.getAttribute('data-index');
-                const removeAssociatesModal = new bootstrap.Modal(document.getElementById('removeAssociatesModal'));
-                const associatesList = document.getElementById('associatesList');
-                associatesList.innerHTML = '';
-
-                areasData[areaIndex].associates.forEach((associate, associateIndex) => {
-                    const listItem = document.createElement('li');
-                    listItem.className = 'list-group-item d-flex justify-content-between align-items-center';
-                    listItem.innerHTML = `
-                        ${associate.name}
-                        <i class="bi bi-trash remove-associate-icon" data-area-index="${areaIndex}" data-associate-index="${associateIndex}"></i>
-                    `;
-                    associatesList.appendChild(listItem);
-                });
-
-                document.querySelectorAll('.remove-associate-icon').forEach(icon => {
-                    icon.addEventListener('click', function() {
-                        const areaIndex = this.getAttribute('data-area-index');
-                        const associateIndex = this.getAttribute('data-associate-index');
-                        showDeleteConfirmationModal('Are you sure you want to delete this associate?', () => {
-                            areasData[areaIndex].associates.splice(associateIndex, 1);
-                            localStorage.setItem('areasData', JSON.stringify(areasData));
-                            renderAreas();
-                            // Remove the list item from the modal
-                            this.parentElement.remove();
-                        });
-                    });
-                });
-
-                removeAssociatesModal.show();
-            });
+            // Initialize area buttons
+            initializeAreaButtons(newArea, index, areaContent);
 
             // Add drag and drop event listeners to the area content
             areaContent.addEventListener('dragover', handleAssociateDragOver);
