+++ index.html
@@ -363,6 +363,7 @@
     <script src="startSessionButton.js"></script>
     <script src="eventListeners.js"></script>
     <script src="areaEventListeners.js"></script>
+    <script src="areaButtons.js"></script>
     <script src="main.js"></script>
     <script>
         // Initialize tooltips
